+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Split out from [pureharm](https://github.com/busymachines/pureharm) as of version `0.0.7`.
 
 - cross compiled to Scala 2.13 -- pending support for scala 3.0.0-RC1
+- add `(??? : Meta[T]).sprout/sproutRefined` ops to easily create `Meta` for sprout new-types.
 
 :warning: Breaking changes :warning:
 - rename `PhantomTypeMetas` to `SproutMetas`